requirements.txt
APScheduler==3.7.0
click==8.0.1
configparser==5.0.2
Flask==2.0.1
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.3.2
pytz==2021.1
six==1.16.0
typing-extensions==3.10.0.2
tzlocal==2.1
Werkzeug==2.0.1
zipp==3.5.0
